Default install and upgrade to "Y" when no answer is given (#1673)
* Default install to "Y" when no answer is given
* Add CHANGELOG
* Add comment to approval prompt
* Default upgrade to "Y" when no answer is given
* Fix underline in upgrade prompt

+++ cli/helm/install.go
@@ -91,7 +91,7 @@ func InstallHelmRelease(options *InstallOptions) error {
 
 	if !options.AutoApprove {
 		confirmation, err := options.UI.Input(&terminal.Input{
-			Prompt: "Proceed with installation? (y/N)",
+			Prompt: "Proceed with installation? (Y/n)",
 			Style:  terminal.InfoStyle,
 			Secret: false,
 		})
@@ -99,7 +99,8 @@ func InstallHelmRelease(options *InstallOptions) error {
 		if err != nil {
 			return err
 		}
-		if common.Abort(confirmation) {
+		// The install will proceed if the user presses enter or responds with "y"/"yes" (case-insensitive).
+		if confirmation != "" && common.Abort(confirmation) {
 			options.UI.Output("Install aborted. Use the command `consul-k8s install -help` to learn how to customize your installation.",
 				terminal.WithInfoStyle())
 			return err
